Move functions from SeasonalError and MSIS out of class
